backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0
supabase==2.10.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
